feat(api): creating folders for gis-shapefiles for exising projects (idas-546)
- moved delete controller to gis-shapefiles-folder.controller file
- deleted gis-shapefiles-folder.delete.controller.js file
- refactored migration.routes.js to update the controller imports
- updated readme to reflect the changes
